Fixes after compiling with GCC -Wextra

+++ efm2riot/static/cpu/efm32_common/periph/timer.c
@@ -109,7 +109,7 @@ int timer_set_absolute(tim_t dev, int channel, unsigned int value)
 {
     TIMER_TypeDef *tim;
 
-    if (channel < 0 || channel >= CC_CHANNELS) {
+    if (channel < 0 || channel >= (int) CC_CHANNELS) {
         return -1;
     }
 
@@ -155,7 +155,7 @@ void TIMER_0_ISR(void)
 {
     TIMER_TypeDef *tim = timer_config[0].timer.dev;
 
-    for (int i = 0; i < CC_CHANNELS; i++) {
+    for (int i = 0; i < (int) CC_CHANNELS; i++) {
         if (tim->IF & (TIMER_IF_CC0 << i)) {
             tim->CC[i].CTRL = _TIMER_CC_CTRL_MODE_OFF;
             tim->IFC = (TIMER_IFC_CC0 << i);
